Update the scroll bar when scrolling the document. Fixes #191.
Plumb `- controller:setScrollbarThumbValue:proportion:identifier:data:`
through to `vimView`.

Note that when you jump to the bottom of the document using `G` the
scrollbar indicator is not positioned at the bottom of the scrollbar.
This is because it is possible to scroll past the bottom of the document
so that the bottom line is at the top of the MMVimView. This behavior is
inherited from MacVim.