Support shift modifier on special keys
The NeoVimView.vimModifierFlags method only checks for the control, option, and command modifiers, which leads to the shift modifier getting lost.
This change mitigates that by checking for the shift modifier and translating it to the `S-` key prefix.

This change has the possible side effect of requiring additional `S-` prefixes, depending on the keyboard layout.
For example, take a US keyboard with a <kbd>-</kbd>/<kbd>_</kbd> key and press that key with control and shift depressed.
Before this change, VimR would transmit `<C-_>`, while after this change it transmits `<C-S-_>`.