Add (Shift-)Tab as a special key
Pressing Shift+Tab generates an NSEvent with characters `"\u{19}"` (and the shift modifier), which the backend doesn’t recognise as the Tab key.
(Instead it gets mapped as <C-Y>, which makes sense given how ASCII control characters are interpreted.)
This change fixes that by adding a corresponding entry to the KeyUtils.specialKeys dict.

Of course sending <C-Y> still works because it is transmitted as `<C-y>`.